refactor: add new data type for money, store it as long

+++ Assets/Scripts/Domain/Economy/ItemType/MoneyItem.cs
@@ -34,9 +34,9 @@ namespace Economy.ItemType
         }
 
         public int MaxItemsToConsume { get { return int.MaxValue; } }
-        public int MaxItemsToWithdraw { get { return _playerResources.Money; } }
+        public int MaxItemsToWithdraw { get { return (int)_playerResources.Money; } }
 
         private readonly PlayerResources _playerResources;
         private readonly ILocalization _localization;
-    }
+	}
 }
