+++ Assets/Scripts/Gui/Craft/UnlockSlotPanel.cs
@@ -67,8 +67,8 @@ namespace Gui.Craft
             if (!CanUnlock)
                 return;
 
-            _resources.Money -= (int)Price.Credits; // TODO: change to some bigger type and remove conversion
-			_resources.Stars -= (int)Price.Stars;
+            _resources.Money -= Price.Credits;
+			_resources.Stars -= Price.Stars;
             _research.AddResearchPoints(_ship.Model.Faction, -Price.Techs);
             var quality = CraftItemQuality.Improved + _ship.Model.Modifications.Count;
             _ship.Experience = Experience.FromLevel(_ship.Experience.Level - GetLevelCost(quality));
